Database.PostgreSQL.Simple.Dsl

Example of usage

A bit of boilerplate is required

newtype UserId = UserId { getUserId :: Int}
                  deriving (Show, ToField, FromField, Eq, Ord)

data User = User { userId :: UserId
                 , userLogin :: String
                 , userPassword :: ByteString
                 } deriving (Show)

data Role = Role { roleUserId :: UserId, roleName :: ByteString }
                  deriving (Show)

data instance Field User t a where
        UserId'   :: Field User "id" UserId
        UserLogin :: Field User "login" String
        UserPass  :: Field User "passwd" ByteString

instance Table User where
        tableName _ = "users"

instance Selectable User where
        entityParser _ = User <$> fromField UserId'
                              <*> fromField UserLogin
                              <*> fromField UserPass

instance Entity User where
        type EntityId User = UserId
        type EntityIdColumn User = "id"

data instance Field Role t a where
        RoleUserId :: Field Role "user_id" UserId
        RoleName   :: Field Role "role" ByteString

instance Table Role where
        tableName _ = "roles"

instance Selectable Role where
        entityParser _ = Role <$> fromField RoleUserId
                              <*> fromField RoleName

getAllUsers :: Connection -> IO [Whole User]
      getAllUsers c = select c $ fromTable

allUsers :: Query (Whole User)
      allUsers = fromTable

allRoles :: Query (Whole Role)
      allRoles = fromTable

This sql snippet maps to next code example

      SELECT roles.role, role.user_id
      FROM users
      INNER JOIN roles ON users.id = roles.user_id
      WHERE user.login = ?

      userRoles :: String -> Query (ByteString, UserId)
      userRoles login =
        innerJoin (\(usr :. rol) -> usr~>UserId' ==. rol~>RoleUserId)
                  -- ^ joining on user.id == roles.user_id
                  (allUsers & where_ (u -> u~>UserLogin ==. val login))
                  -- ^ restrict to user with needed login
                  allRoles
                  -- ^ select all roles
        & project (\(usr :. rol) -> (rol~>RoleName, rol~>RoleUserId))
          -- select only roles.role and role.user_id fields

     SELECT users.*, roles.*
     FROM users, roles
     WHERE users.id = role.user_id

      userRoles2 = crossJoin fromTable fromTable
                 & where_ (\(u:.rol) -> u~>UserId' ==. rol~>RoleUserId)

select :: (IsExpr a, FromRow a) => Connection -> Query a -> IO [a]

Execute query

formatQuery :: IsExpr a => Connection -> Query a -> IO ByteString

fromTable :: forall a. (Table a, Selectable a) => Query (Whole a)

Start query with table - SELECT table.*

with :: (IsExpr with, IsExpr a) => Query with -> Query a -> Query (a :. with)

Adds non recursive WITH query

innerJoin :: ((AsExpr a :. AsExpr b) -> Expr Bool) -> Query a -> Query b -> Query (a :. b)

Inner join on bool expression. Joins WHERE clauses using AND drops ORDER BY, LIMIT, OFFSET.

crossJoin :: Query a -> Query b -> Query (a :. b)

Do a cross join, SELECT .. FROM a,b

where_ :: (AsExpr a -> Expr Bool) -> Query a -> Query a

Append expression to WHERE clause

project :: (AsExpr a -> AsExpr b) -> Query a -> Query b

Select only subset of data

orderBy :: (AsExpr a -> [Sorting]) -> Query a -> Query a

Replace ORDER BY clause

orderOn :: (AsExpr a -> Sorting) -> Query a -> Query a

append to ORDER BY clase

limitTo :: Int -> Query a -> Query a

set LIMIT

offsetTo :: Int -> Query a -> Query a

set OFFSET

val :: ToField a => a -> Expr a

lift value to Expr

whole :: Expr a -> Expr (Whole a)

Select whole entity when projecting

only :: Expr a -> Expr (Only a)

Wrap in Only

data Whole a

Wrapper for selecting whole entity

newtype Only a

A single-value "collection".

This is useful if you need to supply a single parameter to a SQL query, or extract a single column from a SQL result.

Parameter example:

query c "select x from scores where x > ?" (Only (42::Int))

Result example:

xs <- query_ c "select id from users"
forM_ xs $ \(Only id) -> {- ... -}

Constructors

Only

Fields

fromOnly :: a

data h :. t

A composite type to parse your custom data structures without having to define dummy newtype wrappers every time.

 instance FromRow MyData where ...

 instance FromRow MyData2 where ...

then I can do the following for free:

 res <- query' c ...
 forM res $ \(MyData{..} :. MyData2{..}) -> do
   ....

Constructors

h :. t

data Query a

Selection query

data Expr a

Adds phantom type to RawExpr

(==.) :: Expr a -> Expr a -> Expr Bool

(<.) :: Expr a -> Expr a -> Expr Bool

(<=.) :: Expr a -> Expr a -> Expr Bool

(>.) :: Expr a -> Expr a -> Expr Bool

(>=.) :: Expr a -> Expr a -> Expr Bool

(||.) :: Expr Bool -> Expr Bool -> Expr Bool

(&&.) :: Expr Bool -> Expr Bool -> Expr Bool

(~>) :: SingI t => Expr (Whole a) -> Field a t b -> Expr b

data family Field v t b :: *

Data family describing columns

class Table v where

Class for naming tables

Methods

tableName :: Proxy v -> ByteString

class Selectable v where

Class for entities which have columns

Methods

entityParser :: Proxy v -> EntityParser v v

fromField :: (SingI f, FromField a) => Field v f a -> EntityParser v a

Parse named field

class FromExpr (AsExpr a) ~ a => IsExpr a

data EntityParser v a

Parser for entities with columns

entityRowParser :: EntityParser v a -> RowParser a