Database.PostgreSQL.Simple.Dsl

Example of usage

A bit of boilerplate is required

newtype UserId = UserId { getUserId :: Int}
                  deriving (Show, ToField, FromField, Eq, Ord)

data User = User { userId :: UserId
                 , userLogin :: String
                 , userPassword :: ByteString
                 } deriving (Show)

data Role = Role { roleUserId :: UserId, roleName :: ByteString }
                  deriving (Show)
      instance Record User where
        data Field User t a where
          UserKey   :: Field User "id" UserId
          UserLogin :: Field User "login" String
          UserPass  :: Field User "passwd" ByteString
          UserIsDeleted :: Field User "is_deleted" Bool
        recordParser _ = User <$> fromField UserId'
                              <*> fromField UserLogin
                              <*> fromField UserPass
                              <* takeField UserIsDelete

Note that UserIsDeleted fieds is not used in parsers, but is still required to mention

      instance Table User where
        tableName _ = "users"

instance Entity User where
        type EntityId User = UserId
        idField w = w~>UserKey

instance Record Role where
        data Field Role t a where
          RoleUserId :: Field Role "user_id" UserId
          RoleName   :: Field Role "role" ByteString
        recordParser _ = Role <$> fromField RoleUserId
                              <*> fromField RoleName

roles = table "roles" :: From (Rel Role)

allUsers2 = fromTable :: Query (Rel User)
      allRoles = from roles

This sql snippet maps to next code example

   SELECT * FROM users WHERE id = $1

    getUser :: UserId -> Query (Rel User)
    getUser uid = do
      u <- fromTable
      where_ $  u~>UserKey ==. val uid
      return u

      SELECT roles.role, role.user_id
      FROM users, roles
      WHERE users.id = roles.user_id AND user.login = $1

      userRoles :: String -> Query (Rel User :. Rel Role)
      userRoles login = do
          u <- fromTable
          r <- from roles
          where $ u~>UserKey ==. r~>RoleUserId &&. u~>UserLogin ==. val login
          return (u:.r)

With explicit joins. Not recommended for postgres, unless you join more that 6 tables and optimizer is too slow.

      SELECT roles.role, role.user_id
      FROM users INNER JOIN roles ON users.id = roles.user_id
      WHERE user.login = $1

     userRoles2 login = do
       r@(u:.r) <- from $ innerJoin roles recordTable
                        $ \r u -> r~>RoleUserId ==. u~>UserKey
       where $ u~>UserLogin ==. val login
       return r

class Record v where

Class for entities which have columns

Associated Types

data Field v t b :: *

Methods

recordParser :: Proxy v -> RecordParser v v

takeField :: (SingI f, FromField a) => Field v f a -> RecordParser v a

Parse named field

data RecordParser v a

Parser for entities with columns

recordRowParser :: RecordParser v a -> RowParser a

Exequting queries

data Query a

query :: (IsExpr a, FromRow (FromExpr a)) => Connection -> Query a -> IO [FromExpr a]

Execute query and get results

execute :: Connection -> Query r -> IO Int64

Execute discarding results, returns number of rows modified

formatQuery :: IsExpr a => Connection -> Query a -> IO ByteString

Format query for previewing

From

class Record v => Table v where

Class for naming tables

Methods

tableName :: Proxy v -> ByteString

data From a

from :: IsQuery m => From a -> m a

table :: Record a => ByteString -> From (Rel a)

Name a table, add proper type annotation for type safety. This is helpfyl if you can select your record from several tables or views. Using Table class is safer

fromTable :: forall a. (Table a, Record a) => Query (Rel a)

Start query with table - SELECT table.*

innerJoin :: forall a b. (IsExpr a, IsExpr b, Table b) => From a -> From b -> (a -> b -> Expr Bool) -> From (a :. b)

Inner join

crossJoin :: forall a b. (IsExpr a, IsExpr b, Table b) => From a -> From b -> From (a :. b)

Cross join

subSelect :: IsExpr a => Query a -> Query a

Treat query as subquery when joining x <- fromTable y <- subSelect q will turn into SELECT .. FROM x, (SELECT .. from y)

With

with :: (IsQuery m, IsExpr a, IsExpr b) => Query a -> (a -> m b) -> m b

withUpdate :: (IsQuery m, IsExpr a, IsExpr b, IsQuery m) => Update a -> (a -> m b) -> m b

Query

where_ :: IsQuery m => Expr Bool -> m ()

orderBy :: Sorting -> Query ()

append to ORDER BY clase

ascendOn :: Expr a -> Sorting

ascend on expression

descendOn :: Expr t -> Sorting

descend on expression

limitTo :: Int -> Query ()

set LIMIT

offsetTo :: Int -> Query ()

set OFFSET

Updates

data Update a

data UpdExpr a

data Updating r a

update :: From (Rel a) -> (Rel a -> Updating a b) -> Update b

updateTable :: forall a b. Table a => (Rel a -> Updating a b) -> Update b

data Inserting r a

insert :: From (Rel a) -> (Rel a -> Inserting a b) -> Update b

insertIntoTable :: forall a b. Table a => (Rel a -> Inserting a b) -> Update b

data Deleting t a

delete :: From (Rel a) -> (Rel a -> Deleting a b) -> Update b

deleteFromTable :: forall a b. Table a => (Rel a -> Deleting a b) -> Update b

set :: SingI t => Field v t a -> Expr a -> UpdExpr v

(=.) :: forall v t a. SingI t => Field v t a -> Expr a -> UpdExpr v

(=.!) :: (SingI t, ToField a) => Field v t a -> a -> UpdExpr v

setField :: (SingI t, IsUpdate m) => Field v t a -> Expr a -> m v ()

setFields :: IsUpdate m => UpdExpr t -> m t ()

Executing updates

queryUpdate :: (FromExpr t ~ r, IsExpr t, FromRow r) => Connection -> Update t -> IO [r]

executeUpdate :: Connection -> Update a -> IO Int64

formatUpdate :: IsExpr t => Connection -> Update t -> IO ByteString

Expressions

data Rel r

(~>) :: SingI t => Rel a -> Field a t b -> Expr b

Access field of relation

data Expr a

Adds phantom type to RawExpr

val :: ToField a => a -> Expr a

lift value to Expr

(==.) :: Expr a -> Expr a -> Expr Bool

(<.) :: Expr a -> Expr a -> Expr Bool

(<=.) :: Expr a -> Expr a -> Expr Bool

(>.) :: Expr a -> Expr a -> Expr Bool

(>=.) :: Expr a -> Expr a -> Expr Bool

(||.) :: Expr Bool -> Expr Bool -> Expr Bool

(&&.) :: Expr Bool -> Expr Bool -> Expr Bool

(~.) :: Expr Text -> Expr Text -> Expr Bool

like operator

true :: Expr Bool

false :: Expr Bool

isNull :: Expr a -> Expr Bool

isInList :: ToField a => Expr a -> [a] -> Expr Bool

whole :: Expr a -> Expr (Whole a)

Select whole entity when projecting

only :: Expr a -> Expr (Only a)

Wrap in Only

Helpers

newtype Whole a

Wrapper for selecting whole record

Constructors

Whole

Fields

getWhole :: a

newtype Only a

A single-value "collection".

This is useful if you need to supply a single parameter to a SQL query, or extract a single column from a SQL result.

Parameter example:

query c "select x from scores where x > ?" (Only (42::Int))

Result example:

xs <- query_ c "select id from users"
forM_ xs $ \(Only id) -> {- ... -}

Constructors

Only

Fields

fromOnly :: a

data h :. t

A composite type to parse your custom data structures without having to define dummy newtype wrappers every time.

 instance FromRow MyData where ...

 instance FromRow MyData2 where ...

then I can do the following for free:

 res <- query' c ...
 forM res $ \(MyData{..} :. MyData2{..}) -> do
   ....

Constructors

h :. t

Helpers

class IsExpr a

Associated Types

type FromExpr a :: *

Functions

data Function a

function :: ByteString -> Function a

arg :: Expr t -> Function a -> Function a

call :: Function a -> Expr a